fixed but with NOP writing data to R0 during forwarding, added a way to disable forwarding using #define, corrected bug with pip_counter adding to register number, added extra counters

+++ src/main.cpp
@@ -13,7 +13,7 @@ int main()
     // clear terminal
     system("clear");
     Instruction a;
-    signed int N_inst;
+    signed long N_inst;
     short loop_input = 0;
     string line;
     ifstream myfile ("image.txt");
